Send SubmissionRequests
The addin is now capable of sending a SubmissionRequest object to the server through a web api call. We anticipated the structure of the classes of the objects sent to the web api had to be equivalent to the structure of the classes on the server side, however we learned the namespaces containining the classes had to be equivalent for the server to properly serialize the sent objects.